Linked HTML form to calculator function.
Main page will now load up a 'GET' form by default. When an equation is given in the calculate box, a 'POST' form is sent and the equation string is put into the calcEval function. After checking to see if the string is legit and can be calculated, the answer is returned (on a separate page).